+++ test/background_test.rb
@@ -4,12 +4,14 @@ require_relative "test_helper"
 require_relative "interactive_controls"
 require_relative "native_provider_replay"
 require_relative "prompt_cache"
+require_relative "budget_completion"
 require "timeout"
 
 class BackgroundTest < Minitest::Test
   include InteractiveControls
   include NativeProviderReplay
   include PromptCache
+  include BudgetCompletion
   class BlockingClient < FakeClient
     attr_reader :entered, :release